requirements.txt
# Core dependencies
opencv-contrib-python==4.8.1.78
numpy>=1.21.0
pyyaml>=6.0

# Optional (for better performance)
numba>=0.56.0

# Optional (for 3D visualization)
open3d>=0.16.0

# Optional (for GUI)
PyQt5>=5.15.0
pyqtgraph>=0.12.0